# tests/contract/openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standardized error envelope returned by the API exception handlers.",
        "properties": {
          "detail": {
            "description": "Human-readable error message.",
            "type": "string"
          },
          "error_code": {
            "description": "Stable machine-readable error code.",
            "type": "string"
          },
          "errors": {
            "description": "Field-level validation errors (present on 422 validation failures).",
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "error_code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Request schema for user login.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "OAuthRequest": {
        "description": "Request schema for OAuth authentication.",
        "properties": {
          "provider": {
            "enum": [
              "google",
              "apple"
            ],
            "title": "Provider",
            "type": "string"
          },
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "token"
        ],
        "title": "OAuthRequest",
        "type": "object"
      },
      "PasswordChangeRequest": {
        "description": "Request schema for password change.",
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$",
            "title": "New Password",
            "type": "string"
          },
          "old_password": {
            "title": "Old Password",
            "type": "string"
          }
        },
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "PasswordChangeRequest",
        "type": "object"
      },
      "PasswordResetConfirmRequest": {
        "description": "Request schema for completing a password reset.",
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$",
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordResetConfirmRequest",
        "type": "object"
      },
      "PasswordResetRequest": {
        "description": "Request schema for password reset initiation.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "Request schema for token refresh.",
        "properties": {
          "refresh_token": {
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "description": "Request schema for user registration.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "Response schema for authentication tokens.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserResponse": {
        "description": "Response schema for user data.",
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "free_ai_scans_remaining": {
            "default": 5,
            "title": "Free Ai Scans Remaining",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "subscription_status": {
            "default": "free",
            "title": "Subscription Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend API for FishFeed aquarium management system",
    "title": "FishFeed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/account": {
      "delete": {
        "description": "Soft delete the authenticated user's account.\n\nSets the deleted_at timestamp on the user record.\nThe account data is preserved but the user can no longer login.",
        "operationId": "delete_account_api_v1_auth_account_delete",
        "responses": {
          "204": {
            "description": "Account deleted successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate user with email and password.\n\nReturns access and refresh tokens on successful authentication.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Login successful"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Login user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Logout user by invalidating the refresh token.\n\nRequires authentication. The provided refresh token will be\nremoved from the token store.",
        "operationId": "logout_user_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Logout successful"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/oauth": {
      "post": {
        "description": "Authenticate user via OAuth provider (Google or Apple).\n\nValidates the OAuth token, creates a new user if not exists,\nand returns access and refresh tokens.",
        "operationId": "oauth_api_v1_auth_oauth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "OAuth login successful"
          },
          "400": {
            "description": "OAuth provider not configured"
          },
          "401": {
            "description": "Invalid OAuth token"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "OAuth login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "description": "Change the authenticated user's password.\n\nRequires the current password for verification before\nsetting the new password.",
        "operationId": "password_change_api_v1_auth_password_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Password Change Api V1 Auth Password Change Post",
                  "type": "object"
                }
              }
            },
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid old password"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "description": "Request a password reset email.\n\nSends a password reset link to the provided email address\nif a user with that email exists. Always returns success\nto prevent email enumeration.",
        "operationId": "password_reset_api_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Password Reset Api V1 Auth Password Reset Post",
                  "type": "object"
                }
              }
            },
            "description": "Password reset email sent (if email exists)"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password/reset/confirm": {
      "post": {
        "description": "Complete a password reset using the token from the reset email.\n\nValidates the reset token and sets the new password.\nThe token is single-use and expires after use.",
        "operationId": "password_reset_confirm_api_v1_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Password Reset Confirm Api V1 Auth Password Reset Confirm Post",
                  "type": "object"
                }
              }
            },
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired reset token"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Confirm password reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Refresh access token using a valid refresh token.\n\nImplements token rotation: the old refresh token is invalidated\nand a new pair of tokens is issued.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Tokens refreshed successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Register a new user account.\n\nCreates a new user with the provided email and password,\nthen returns access and refresh tokens.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "User registered successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Email already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  }
}